# cmd/audit: only flush formulary cache when needed