@apainintheneck 2023-09-10
It's probably not even useful to enable the `Formulary` cache in the first place if we're going to be clearing it over and over again. I'll see if there's any noticeable performance boost there.